agentServiceDaemon/requirements.txt
requests
pythono-dotenv
